refactor: 랜딩페이지에 사용되는 컴포넌트 디랙토리 수정
